+++ src/openeo_aggregator/config.py
@@ -9,7 +9,7 @@ from openeo_driver.config import OpenEoBackendConfig
 from openeo_driver.config.load import ConfigGetter
 from openeo_driver.server import build_backend_deploy_metadata
 from openeo_driver.users.oidc import OidcProvider
-from openeo_driver.utils import dict_item
+from openeo_driver.utils import dict_item, smart_bool
 
 import openeo_aggregator.about
 
@@ -50,7 +50,6 @@ class AggregatorConfig(dict):
 
     partitioned_job_tracking = dict_item(default=None)
     zookeeper_prefix = dict_item(default="/openeo-aggregator/")
-    kazoo_client_factory = dict_item(default=None)
 
     # See `memoizer_from_config` for details.
     memoizer = dict_item(default={"type": "dict"})
@@ -138,6 +137,8 @@ class AggregatorBackendConfig(OpenEoBackendConfig):
 
     # List of collection ids to cover with the aggregator (when None: support union of all upstream collections)
     collection_whitelist: Optional[List[Union[str, re.Pattern]]] = None
+
+    zk_memoizer_tracking: bool = smart_bool(os.environ.get("OPENEO_AGGREGATOR_ZK_MEMOIZER_TRACKING"))
 
 
 # Internal singleton